Pre release 1.0

+++ ocsreports/tele_actives.php
@@ -8,10 +8,12 @@
 // code is always made freely available.
 // Please refer to the General Public Licence http://www.gnu.org/ or Licence.txt
 //====================================================================================
-//Modified on $Date: 2006-12-21 18:13:47 $$Author: plemmet $($Revision: 1.4 $)
+//Modified on $Date: 2007-01-26 17:05:42 $$Author: plemmet $($Revision: 1.5 $)
 
 if( isset($_GET["suppack"])) {
-	@mysql_query("DELETE FROM download_enable WHERE ID=".$_GET["suppack"], $_SESSION["writeServer"]) or die(mysql_error());	
+	
+	if( !isset($_GET["nonnot"]) )
+		@mysql_query("DELETE FROM download_enable WHERE ID=".$_GET["suppack"], $_SESSION["writeServer"]) or die(mysql_error());	
 	
 	$reqSupp = "DELETE FROM devices WHERE name='DOWNLOAD' AND ivalue=".$_GET["suppack"];
 	if( isset($_GET["nonnot"]) )
@@ -24,14 +26,14 @@ $canAc = 2;
 PrintEnTete( $l->g(481) );	
 	
 $lbl = "pack";	
-$sql = "d.fileid = e.fileid";
-$whereId = "d.FILEID=e.FILEID AND d.FILEID";
-$linkId = "d.FILEID";
-$select = array("ID"=>$l->g(460), "d.FILEID"=>$l->g(475), "NAME"=>$l->g(49), 
+$sql = "";
+$whereId = "e.FILEID";
+$linkId = "e.FILEID";
+$select = array("ID"=>$l->g(460), "e.FILEID"=>$l->g(475), "NAME"=>$l->g(49), 
 "PRIORITY"=>$l->g(440),"INFO_LOC"=>$l->g(470), "PACK_LOC"=>$l->g(471), 
 "FRAGMENTS"=>$l->g(480), "SIZE"=>$l->g(462), "OSNAME"=>$l->g(25));	
-$selectPrelim = array("d.FILEID"=>"d.FILEID");	
-$from = "download_available d, download_enable e";
+$selectPrelim = array("e.FILEID"=>"e.FILEID", "d.name"=>"Nom");		
+$from = "download_enable e LEFT JOIN download_available d ON d.fileid = e.fileid";
 $fromPrelim = "";
 $group = "";
 $order = "";
